requirements.txt
requests~=2.32
PyYAML~=6.0
python-dateutil~=2.8
tqdm~=4.66
scikit-learn~=1.3
numpy~=1.26
beautifulsoup4~=4.12
# 可選：若要啟用 Gate.io BigData 經由 Playwright 的 fallback，請安裝並執行 `playwright install chromium`
playwright~=1.48
